+++ src/main/kotlin/site/siredvin/turtlematic/computercraft/ComputerCraftProxy.kt
@@ -12,14 +12,8 @@ import site.siredvin.peripheralium.computercraft.turtle.PeripheralTurtleUpgrade
 import site.siredvin.peripheralium.util.ItemUtil
 import site.siredvin.turtlematic.Turtlematic
 import site.siredvin.turtlematic.common.setup.Items
-import site.siredvin.turtlematic.computercraft.peripheral.automatas.AutomataCorePeripheral
-import site.siredvin.turtlematic.computercraft.peripheral.automatas.EndAutomataCorePeripheral
-import site.siredvin.turtlematic.computercraft.peripheral.automatas.EnormousAutomataCorePeripheral
-import site.siredvin.turtlematic.computercraft.peripheral.automatas.HusbandryAutomataCorePeripheral
-import site.siredvin.turtlematic.computercraft.peripheral.forged.BrewingAutomataCorePeripheral
-import site.siredvin.turtlematic.computercraft.peripheral.forged.EnchantingAutomataCorePeripheral
-import site.siredvin.turtlematic.computercraft.peripheral.forged.MasonAutomataCorePeripheral
-import site.siredvin.turtlematic.computercraft.peripheral.forged.SmithingAutomataCorePeripheral
+import site.siredvin.turtlematic.computercraft.peripheral.automatas.*
+import site.siredvin.turtlematic.computercraft.peripheral.forged.*
 import site.siredvin.turtlematic.computercraft.peripheral.misc.*
 import site.siredvin.turtlematic.computercraft.turtle.BlockTurtleUpgrade
 import site.siredvin.turtlematic.computercraft.turtle.ClockwiseAnimatedTurtleUpgrade
@@ -66,14 +60,17 @@ object ComputerCraftProxy {
         TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.MASON_AUTOMATA_CORE) { turtle, side, tier -> MasonAutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.ENCHANTING_AUTOMATA_CORE) { turtle, side, tier -> EnchantingAutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.SMITHING_AUTOMATA_CORE) { turtle, side, tier -> SmithingAutomataCorePeripheral(turtle, side, tier) })
+        TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.MERCANTILE_AUTOMATA_CORE) { turtle, side, tier -> MercantileAutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(StarboundTurtleUpgrade.dynamic(Items.STARBOUND_BREWING_AUTOMATA_CORE) { turtle, side, tier -> BrewingAutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(StarboundTurtleUpgrade.dynamic(Items.STARBOUND_MASON_AUTOMATA_CORE) { turtle, side, tier -> MasonAutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(StarboundTurtleUpgrade.dynamic(Items.STARBOUND_ENCHANTING_AUTOMATA_CORE) { turtle, side, tier -> EnchantingAutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(StarboundTurtleUpgrade.dynamic(Items.STARBOUND_SMITHING_AUTOMATA_CORE) { turtle, side, tier -> SmithingAutomataCorePeripheral(turtle, side, tier) })
+        TURTLE_UPGRADES.add(StarboundTurtleUpgrade.dynamic(Items.STARBOUND_MERCANTILE_AUTOMATA_CORE) { turtle, side, tier -> MercantileAutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.CREATIVE_BREWING_AUTOMATA_CORE) { turtle, side, tier -> BrewingAutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.CREATIVE_MASON_AUTOMATA_CORE) { turtle, side, tier -> MasonAutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.CREATIVE_ENCHANTING_AUTOMATA_CORE) { turtle, side, tier -> EnchantingAutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.CREATIVE_SMITHING_AUTOMATA_CORE) { turtle, side, tier -> SmithingAutomataCorePeripheral(turtle, side, tier) })
+        TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.CREATIVE_MERCANTILE_AUTOMATA_CORE) { turtle, side, tier -> MercantileAutomataCorePeripheral(turtle, side, tier) })
 
         TURTLE_UPGRADES.add(PeripheralTurtleUpgrade.dynamic(Items.SOUL_SCRAPPER) { turtle, side -> SoulScrapperPeripheral(TurtlePeripheralOwner(turtle, side)) })
         TURTLE_UPGRADES.add(PeripheralTurtleUpgrade.dynamic(net.minecraft.world.item.Items.LAVA_BUCKET, { turtle, side -> LavaBucketPeripheral(TurtlePeripheralOwner(turtle, side)) }, WITH_TURTLEMATIC_ID))